update the home page to be the welcome page. updated the dashboard to be able to fit the menu items. updated snapshots.
