--- imports/pages/celebrate/message-from-shane/index.js
@@ -1,129 +1,0 @@
-// @flow
-import { Link } from "react-router";
-// import { Leaves } from "../components/layout";
-// import { SolidLeaf, StripedLeaf } from "../components/leaves";
-import Meta from "../../../components/shared/meta";
-import Video from "../../../components/@primitives/players/video";
-
-/* eslint-disable max-len */
-export const Shane = () => (
-  <div>
-    <Meta title="Message From Shane" />
-    {/* Header */}
-    <style>{".oo-player-container { min-width: 0 !important }"}</style>
-    <div className="push-big-sides@lap-and-up soft@lap-and-up soft-top@lap-and-up">
-      <div className="soft soft-bottom soft-top@lap-and-up">
-        <div className="soft-bottom">
-          <Video id={"VoNTB1ODE66MXxAFqcCqRb04c2WT9biy"} autoplay={false} />
-        </div>
-        <h1 className={`text-center soft-top soft-bottom`}>
-          Welcome to NewSpring&#39;s 2017 Annual Report!
-        </h1>
-        <p style={{ fontFamily: "colfax-web" }}>
-          The annual report is an opportunity to reflect on what God has done
-          and to see why the local church is the best investment we can make.
-        </p>
-        <p style={{ fontFamily: "colfax-web" }}>
-          At NewSpring, we have so much to celebrate from 2016! Thousands of
-          lives have been changed as a result of your generosity. Throughout
-          this report, you&#39;ll meet people who connected to Jesus and the
-          church because of the way you give, serve, and pray.
-        </p>
-        <p style={{ fontFamily: "colfax-web" }}>
-          You&#39;ll also see that 2016 was a year of transition for our church.
-          Growing things change, and NewSpring is no exception. We saw changes
-          in leadership and at our campuses. Lexington merged with Columbia, and
-          two campuses (Powdersville and Clemson) moved into new permanent
-          buildings. We held our largest Gauntlet ever, and we launched Connect
-          - an event designed to help NewSpring attenders connect to Jesus and
-          each other.
-        </p>
-        <p style={{ fontFamily: "colfax-web" }}>
-          In every season of our church&#39;s history, from 2000 to now,
-          we&#39;ve seen Jesus come through on His promise to build His church.
-          My hope is that as you read this report, you will see the miracles God
-          is doing every week at NewSpring, and you&#39;ll join me in
-          anticipating what God has in store from us next.
-        </p>
-        <p style={{ fontFamily: "colfax-web" }}>
-          Shane Duffey<br />Executive Pastor
-        </p>
-      </div>
-    </div>
-    {
-      // XXX: Do we really want this?
-      // <div
-      //   style={{
-      //     position: "absolute",
-      //     top: "800px",
-      //     right: "0",
-      //   }}
-      //   className="visuallyhidden@handheld visuallyhidden@lap one-half display-inline-block"
-      // >
-      //   <Leaves>
-      //     <SolidLeaf
-      //       className="locked-bottom locked-right"
-      //       size="200px"
-      //       color="text-secondary"
-      //       style={{
-      //         transform: "rotate(225deg)",
-      //         marginBottom: "-80px",
-      //         marginRight: "-50px",
-      //       }}
-      //     />
-      //     <SolidLeaf
-      //       className="locked-bottom locked-right"
-      //       size="100px"
-      //       color="text-primary"
-      //       style={{
-      //         transform: "rotate(90deg)",
-      //         marginBottom: "-55px",
-      //         marginRight: "170px",
-      //       }}
-      //     />
-      //     <SolidLeaf
-      //       className="locked-bottom locked-right"
-      //       size="200px"
-      //       color="text-tertiary"
-      //       style={{
-      //         transform: "rotate(0deg)",
-      //         marginBottom: "-170px",
-      //         marginRight: "-10px",
-      //       }}
-      //     />
-      //     <StripedLeaf
-      //       className="locked-bottom locked-right"
-      //       size="120px"
-      //       color="text-secondary"
-      //       style={{
-      //         transform: "rotate(270deg)",
-      //         marginBottom: "-180px",
-      //         marginRight: "150px",
-      //       }}
-      //     />
-      //   </Leaves>
-      // </div>
-    }
-    <div className="background--primary text-center text-light-primary soft">
-      <h3 className="push-double-top">Keep Reading</h3>
-      <Link
-        className="btn--light push-double-bottom"
-        to="/annualreport/finances"
-      >
-        Go to Finance Report
-      </Link>
-    </div>
-  </div>
-);
-
-const Routes = [
-  {
-    path: "home",
-    component: Shane,
-  },
-];
-
-export default {
-  Shane,
-  Routes,
-};
